Refactor arbitrary `NUMERIC` `Postgres` types to `DOUBLE PRECISION`

+++ backend/queries/postgres/create_table_journeys.sql
@@ -7,8 +7,8 @@ CREATE TABLE IF NOT EXISTS journeys (
     return_date TIMESTAMP NOT NULL,
     return_station_id INT NOT NULL,
 
-    distance_meters NUMERIC NOT NULL,
-    duration_seconds NUMERIC NOT NULL,
+    distance_meters DOUBLE PRECISION NOT NULL,
+    duration_seconds DOUBLE PRECISION NOT NULL,
 
     CONSTRAINT fk_departure_station_id FOREIGN KEY(departure_station_id) REFERENCES stations(station_id),
     CONSTRAINT fk_return_station_id FOREIGN KEY(return_station_id) REFERENCES stations(station_id)
